+++ interfaces/HO-Portal/src/app/components/physical-card-popup/physical-card-popup.component.ts
@@ -3,20 +3,19 @@ import { Component, Input, OnInit } from '@angular/core';
 import { AlertController, IonicModule, ModalController } from '@ionic/angular';
 import { TranslateModule, TranslateService } from '@ngx-translate/core';
 import { DateFormat } from 'src/app/enums/date-format.enum';
+import { WalletStatus121 } from '../../../../../../services/121-service/src/payments/fsp-integration/intersolve-visa/enum/wallet-status-121.enum';
 import { AuthService } from '../../auth/auth.service';
 import Permission from '../../auth/permission.enum';
-import {
-  PhysicalCard,
-  PhysicalCardStatus,
-} from '../../models/physical-card.model';
+import { PhysicalCard } from '../../models/physical-card.model';
 import { ErrorHandlerService } from '../../services/error-handler.service';
 import { ProgramsServiceApiService } from '../../services/programs-service-api.service';
 import { actionResult } from '../../shared/action-result';
+import { SharedModule } from '../../shared/shared.module';
 
 @Component({
   selector: 'app-physical-card-popup',
   standalone: true,
-  imports: [CommonModule, IonicModule, TranslateModule],
+  imports: [CommonModule, IonicModule, TranslateModule, SharedModule],
   templateUrl: './physical-card-popup.component.html',
   styleUrls: ['./physical-card-popup.component.scss'],
 })
@@ -25,14 +24,26 @@ export class PhysicalCardPopupComponent implements OnInit {
   public programId: number;
 
   @Input({ required: true })
-  public card: PhysicalCard;
+  public card: PhysicalCard = {
+    status: WalletStatus121.Active,
+  } as PhysicalCard;
 
   @Input({ required: true })
   public currency: string;
 
+  @Input({ required: true })
+  public referenceId: string;
+
+  @Input({ required: true })
+  public showButtons: boolean;
+
   public DateFormat = DateFormat;
-  public PhysicalCardStatus = PhysicalCardStatus;
-  public blockButtonLabel: string;
+  public WalletStatus121 = WalletStatus121;
+
+  public isCardPaused: boolean;
+
+  public issueLoading = false;
+  public pauseLoading = false;
 
   constructor(
     private modalController: ModalController,
@@ -42,57 +53,60 @@ export class PhysicalCardPopupComponent implements OnInit {
     private errorHandlerService: ErrorHandlerService,
     private authService: AuthService,
   ) {}
-
-  ngOnInit() {
-    this.blockButtonLabel = this.getBlockButtonLabel(this.card);
-  }
-
-  private getBlockButtonLabel(card: PhysicalCard) {
-    return card?.status.toUpperCase() === PhysicalCardStatus.blocked
-      ? this.translate.instant(
-          'registration-details.physical-cards-overview.unblock-card',
-        )
-      : this.translate.instant(
-          'registration-details.physical-cards-overview.block-card',
-        );
+  ngOnInit(): void {
+    this.isCardPaused = this.card.status === WalletStatus121.Paused;
   }
 
   public closeModal() {
     this.modalController.dismiss();
   }
 
-  public canBlock() {
+  private canPause() {
     return this.authService.hasPermission(
       this.programId,
       Permission.FspDebitCardBLOCK,
     );
   }
 
-  public canUnblock() {
+  private canUnpause() {
     return this.authService.hasPermission(
       this.programId,
       Permission.FspDebitCardUNBLOCK,
     );
   }
 
-  public canUseButton() {
-    return this.card?.status.toUpperCase() === PhysicalCardStatus.blocked
-      ? this.canUnblock()
+  public canIssueNewCard(): boolean {
+    return this.authService.hasPermission(
+      this.programId,
+      Permission.FspDebitCardCREATE,
+    );
+  }
+
+  public canUsePauseButton() {
+    return this.card.status === WalletStatus121.Paused
+      ? this.canUnpause()
         ? true
         : false
-      : this.canBlock()
+      : this.canPause()
       ? true
       : false;
   }
 
-  toggleBlockButton(card: PhysicalCard) {
-    const block = card.status.toUpperCase() !== PhysicalCardStatus.blocked;
+  togglePauseButton() {
+    this.pauseLoading = true;
+    const block = this.card.status !== WalletStatus121.Paused;
     this.progamsServiceApiService
-      .toggleBlockWallet(this.programId, card.tokenCode, block)
+      .toggleBlockWallet(this.programId, this.card.tokenCode, block)
       .then((response) => {
         let message = '';
         if (response.status === 204) {
-          message = this.translate.instant('common.update-success');
+          message = block
+            ? this.translate.instant(
+                'registration-details.physical-cards-overview.action-result.pause-success',
+              )
+            : this.translate.instant(
+                'registration-details.physical-cards-overview.action-result.unpause-success',
+              );
         } else if (response.status === 405) {
           message = this.translate.instant('common.update-error', {
             error: response.data?.code,
@@ -115,11 +129,45 @@ export class PhysicalCardPopupComponent implements OnInit {
             true,
           );
         }
+      })
+      .finally(() => {
+        this.pauseLoading = false;
       });
   }
 
-  sendReplacementCardButtonClick(card: PhysicalCard) {
-    // TODO: Implement send replacement card functionality
-    console.log('card: ', card);
+  issueNewCardButtonClick() {
+    this.issueLoading = true;
+    this.progamsServiceApiService
+      .issueNewCard(this.programId, this.referenceId)
+      .then(() => {
+        actionResult(
+          this.alertController,
+          this.translate,
+          this.translate.instant(
+            'registration-details.physical-cards-overview.action-result.new-card-success',
+          ),
+          true,
+        );
+      })
+      .catch((error) => {
+        console.log('error: ', error);
+        if (error && error.error) {
+          const errorMessage = this.translate.instant(
+            'registration-details.physical-cards-overview.action-result.new-card-error',
+            {
+              error: error.error.errors,
+            },
+          );
+          actionResult(
+            this.alertController,
+            this.translate,
+            errorMessage,
+            true,
+          );
+        }
+      })
+      .finally(() => {
+        this.issueLoading = false;
+      });
   }
 }