+++ interfaces/HO-Portal/src/app/components/registration-physical-card-overview/registration-physical-card-overview.component.ts
@@ -3,11 +3,10 @@ import { Component, Input, OnInit } from '@angular/core';
 import { IonicModule, ModalController } from '@ionic/angular';
 import { TranslateModule } from '@ngx-translate/core';
 import { Person } from 'src/app/models/person.model';
-import {
-  PhysicalCard,
-  PhysicalCardStatus,
-} from 'src/app/models/physical-card.model';
+import { PhysicalCard } from 'src/app/models/physical-card.model';
 import { ProgramsServiceApiService } from 'src/app/services/programs-service-api.service';
+import { WalletStatus121 } from '../../../../../../services/121-service/src/payments/fsp-integration/intersolve-visa/enum/wallet-status-121.enum';
+import RegistrationStatus from '../../enums/registration-status.enum';
 import { PhysicalCardPopupComponent } from '../physical-card-popup/physical-card-popup.component';
 import { RegistrationPageTableComponent } from '../registration-page-table/registration-page-table.component';
 
@@ -34,8 +33,17 @@ export class RegistrationPhysicalCardOverviewComponent implements OnInit {
   @Input()
   public currency: string;
 
+  @Input()
+  public nrPayments: Person['nrPayments'];
+
+  @Input()
+  public registrationStatus: Person['status'];
+
   public physicalCards: PhysicalCard[];
-  public PhysicalCardStatus = PhysicalCardStatus;
+  public WalletStatus121 = WalletStatus121;
+  public latestCard: PhysicalCard;
+
+  public loading = true;
 
   constructor(
     private programsService: ProgramsServiceApiService,
@@ -43,10 +51,31 @@ export class RegistrationPhysicalCardOverviewComponent implements OnInit {
   ) {}
 
   public async ngOnInit() {
-    this.physicalCards = await this.programsService.getPhysicalCards(
-      this.programId,
-      this.referenceId,
-    );
+    if (!this.nrPayments || Number(this.nrPayments) === 0) {
+      this.loading = false;
+      return;
+    }
+
+    this.physicalCards = (
+      await this.programsService.getPhysicalCards(
+        this.programId,
+        this.referenceId,
+      )
+    ).sort((a, b) => {
+      if (a.issuedDate < b.issuedDate) {
+        return 1;
+      }
+
+      if (a.issuedDate > b.issuedDate) {
+        return -1;
+      }
+
+      return 0;
+    });
+
+    this.latestCard = this.physicalCards[0];
+
+    this.loading = false;
   }
 
   public async openCardDetails(card: PhysicalCard) {
@@ -56,8 +85,25 @@ export class RegistrationPhysicalCardOverviewComponent implements OnInit {
         card,
         currency: this.currency,
         programId: this.programId,
+        referenceId: this.referenceId,
+        showButtons: this.latestCard.tokenCode === card.tokenCode,
       },
     });
     await modal.present();
+  }
+
+  public showPhysicalCardOverview(): boolean {
+    const acceptedStatuses = [
+      RegistrationStatus.included,
+      RegistrationStatus.completed,
+      RegistrationStatus.inclusionEnded,
+      RegistrationStatus.rejected,
+    ];
+
+    if (acceptedStatuses.includes(this.registrationStatus)) {
+      return true;
+    }
+
+    return false;
   }
 }