+++ interfaces/HO-Portal/src/app/services/programs-service-api.service.ts
@@ -76,9 +76,9 @@ export class ProgramsServiceApiService {
   }
 
   deleteRegistrations(programId: number, referenceIds: string[]): Promise<any> {
-    return this.apiService.post(
+    return this.apiService.delete(
       environment.url_121_service_api,
-      `/programs/${programId}/registrations/delete`,
+      `/programs/${programId}/registrations`,
       {
         referenceIds,
       },
@@ -466,6 +466,19 @@ export class ProgramsServiceApiService {
     );
 
     return !!response && !!response.wallets ? response.wallets : [];
+  }
+
+  public async issueNewCard(
+    programId: number,
+    referenceId: string,
+  ): Promise<any> {
+    const res = await this.apiService.put(
+      environment.url_121_service_api,
+      `/programs/${programId}/fsp-integration/intersolve-visa/customers/${referenceId}/wallets`,
+      {},
+    );
+
+    return res;
   }
 
   public async toggleBlockWallet(