+++ services/121-service/src/payments/fsp-integration/intersolve-visa/enum/wallet-status-121.enum.ts
@@ -1,6 +1,6 @@
 export enum WalletStatus121 {
-  Substituted = 'Substituted',
   Active = 'Active',
-  Inactive = 'Inactive',
+  Issued = 'Issued',
   Blocked = 'Blocked',
+  Paused = 'Paused',
 }