+++ services/121-service/src/utils/registration-data-query/registration-data-query.service.ts
@@ -1,0 +1,73 @@
+import { Injectable } from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
+import { Repository, SelectQueryBuilder } from 'typeorm';
+import { v4 as uuid } from 'uuid';
+import {
+  RegistrationDataOptions,
+  RegistrationDataRelation,
+} from '../../registration/dto/registration-data-relation.model';
+import { GenericAttributes } from '../../registration/enum/custom-data-attributes';
+import { RegistrationDataEntity } from '../../registration/registration-data.entity';
+import { RegistrationEntity } from '../../registration/registration.entity';
+
+@Injectable()
+export class RegistrationDataQueryService {
+  @InjectRepository(RegistrationEntity)
+  private readonly registrationRepository: Repository<RegistrationEntity>;
+
+  public async getPaDetails(
+    referenceIds: string[],
+    relationOptions: RegistrationDataOptions[],
+  ): Promise<any> {
+    const query = this.registrationRepository
+      .createQueryBuilder('registration')
+      .select([
+        `registration.referenceId as "referenceId"`,
+        `registration."${GenericAttributes.phoneNumber}"`,
+        `registration."${GenericAttributes.preferredLanguage}"`,
+        `coalesce(registration."${GenericAttributes.paymentAmountMultiplier}",1) as "paymentAmountMultiplier"`,
+      ])
+      .where(`registration.referenceId IN (:...referenceIds)`, {
+        referenceIds,
+      });
+    for (const r of relationOptions) {
+      query.select((subQuery) => {
+        return this.customDataEntrySubQuery(subQuery, r.relation);
+      }, r.name);
+    }
+
+    return await query.getRawMany();
+  }
+
+  public customDataEntrySubQuery(
+    subQuery: SelectQueryBuilder<any>,
+    relation: RegistrationDataRelation,
+  ): SelectQueryBuilder<any> {
+    const uniqueSubQueryId = uuid().replace(/-/g, '').toLowerCase();
+    subQuery = subQuery
+      .where(`"${uniqueSubQueryId}"."registrationId" = registration.id`)
+      .from(RegistrationDataEntity, uniqueSubQueryId);
+    if (relation.programQuestionId) {
+      subQuery = subQuery.andWhere(
+        `"${uniqueSubQueryId}"."programQuestionId" = ${relation.programQuestionId}`,
+      );
+    } else if (relation.monitoringQuestionId) {
+      subQuery = subQuery.andWhere(
+        `"${uniqueSubQueryId}"."monitoringQuestionId" = ${relation.monitoringQuestionId}`,
+      );
+    } else if (relation.programCustomAttributeId) {
+      subQuery = subQuery.andWhere(
+        `"${uniqueSubQueryId}"."programCustomAttributeId" = ${relation.programCustomAttributeId}`,
+      );
+    } else if (relation.fspQuestionId) {
+      subQuery = subQuery.andWhere(
+        `"${uniqueSubQueryId}"."fspQuestionId" = ${relation.fspQuestionId}`,
+      );
+    }
+    // Because of string_agg no distinction between multi-select and other is needed
+    subQuery.addSelect(
+      `string_agg("${uniqueSubQueryId}".value,'|' order by value)`,
+    );
+    return subQuery;
+  }
+}